// src/app/shared/models/empresa/empresa.ts
export class Empresa {
  idempresa: number;
  cnpj: string;
  inscricaoestadual: string;
  razaosocial: string;
  fantasia: string;
  email: string;
  inscricaoisenta = true;
  rntrc: string;
  tipoemitente: string;
  tpTransp: string;
  tpemis: string;

  cep: string;
  uf: string;
  logradouro: string;
  complemento: string;
  pais = 'Brasil';
  municipio: string;
  codigomunicipio: number;
  numero: string;
  bairro: string;
  fone: string;
  observacao: string;
  flagcanalverde = false;

  flagdel = false;
}